% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.R
\name{install_packages}
\alias{install_packages}
\title{Installs necessary packages}
\usage{
install_packages()
}
\value{
No value is returned.
}
\description{
This function checks if the `devtools` package is installed and installs it if not. It then installs the `scmamp` package from GitHub.
}
