% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win-tie-loss.R
\name{random.dataframe}
\alias{random.dataframe}
\title{Generate a Random Data Frame}
\usage{
random.dataframe()
}
\value{
A data frame with random values.
}
\description{
This function generates a random data frame with a predefined number of rows 
and columns, simulating different methods and datasets. The resulting data 
frame is saved as a CSV file.
}
\examples{
random.dataframe()
}
