% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win-tie-loss.R
\name{win.tie.loss.compute}
\alias{win.tie.loss.compute}
\title{Compute Win-Tie-Loss Summary}
\usage{
win.tie.loss.compute(data, measure.type)
}
\arguments{
\item{data}{A data frame containing the scores of different methods.}

\item{measure.type}{An integer indicating the measure type (1 or 2) which 
defines the win/loss criteria.}
}
\value{
A data frame summarizing the number of wins, ties, and losses for each method.
}
\description{
This function calculates the win, tie, and loss summary between pairs of 
methods based on their scores for a given measure type.
}
\examples{
name.file = "~/WinTieLoss/Data/clp.csv"
data = data.frame(read.csv(name.file))
data = data[,-1]
methods.names = colnames(data)

df_res.mes <- wtl.measures()
filtered_res.mes <- filter(df_res.mes, names == "clp")
measure.type = as.numeric(filtered_res.mes$type)

res = win.tie.loss.compute(data = data, measure.type)
res$method <- factor(res$method, levels = methods.names)
res <- res[order(res$method), ]

save = paste(FolderResults, "/clp.csv", sep="")
write.csv(res, save, row.names = FALSE)

wtl = c("win", "tie", "loss")
colnames(res) = wtl 

save = paste(FolderResults, "/clp.pdf", sep="")
win.tie.loss.plot(data = res, 
                  names.methods = methods.names, 
                  name.file = save, width = 18, height = 10, 
                  bottom = 2, left = 11, top = 0, right = 1, 
                  size.font = 2.0,wtl = wtl)


}
