% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win-tie-loss.R
\name{win.tie.loss.plot}
\alias{win.tie.loss.plot}
\title{Plot Win-Tie-Loss Barplot}
\usage{
win.tie.loss.plot(
  data,
  names.methods,
  name.file,
  width,
  height,
  bottom,
  left,
  top,
  right,
  size.font,
  wtl
)
}
\arguments{
\item{data}{A data frame containing the win-tie-loss summary.}

\item{names.methods}{A vector of method names to be displayed in the plot.}

\item{name.file}{The name of the output PDF file.}

\item{width}{The width of the plot in inches.}

\item{height}{The height of the plot in inches.}

\item{bottom}{Margin size for the bottom of the plot.}

\item{left}{Margin size for the left of the plot.}

\item{top}{Margin size for the top of the plot.}

\item{right}{Margin size for the right of the plot.}

\item{size.font}{Font size for labels and axis text.}

\item{wtl}{A character vector indicating the labels for win, tie, and loss.}
}
\value{
A PDF file containing the win-tie-loss barplot.
}
\description{
This function generates a barplot visualizing the win-tie-loss summary 
for different methods. The plot is saved as a PDF file.
}
\examples{
#' 
result <- win.tie.loss.compute(df, 1)
win.tie.loss.plot(result, names(result$method), "win_tie_loss.pdf", 
                  10, 7, 5, 4, 3, 2, 1.5, c("Win", "Tie", "Loss"))
}
