% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win-tie-loss.R
\name{win.tie.loss.plot}
\alias{win.tie.loss.plot}
\title{Plot Win-Tie-Loss Barplot}
\usage{
win.tie.loss.plot(
  data,
  names.methods,
  name.file,
  width,
  height,
  bottom,
  left,
  top,
  right,
  size.font,
  wtl
)
}
\arguments{
\item{data}{A data frame containing the win-tie-loss summary.}

\item{names.methods}{A vector of method names to be displayed in the plot.}

\item{name.file}{The name of the output PDF file.}

\item{width}{The width of the plot in inches.}

\item{height}{The height of the plot in inches.}

\item{bottom}{Margin size for the bottom of the plot.}

\item{left}{Margin size for the left of the plot.}

\item{top}{Margin size for the top of the plot.}

\item{right}{Margin size for the right of the plot.}

\item{size.font}{Font size for labels and axis text.}

\item{wtl}{A character vector indicating the labels for win, tie, and loss.}
}
\value{
A PDF file containing the win-tie-loss barplot.
}
\description{
This function generates a barplot visualizing the win-tie-loss summary 
for different methods. The plot is saved as a PDF file.
}
\examples{
name.file = "~/WinTieLoss/Data/clp.csv"
data = data.frame(read.csv(name.file))
data = data[,-1]
methods.names = colnames(data)

df_res.mes <- wtl.measures()
filtered_res.mes <- filter(df_res.mes, names == "clp")
measure.type = as.numeric(filtered_res.mes$type)

res = win.tie.loss.compute(data = data, measure.type)
res$method <- factor(res$method, levels = methods.names)
res <- res[order(res$method), ]

save = paste(FolderResults, "/clp.csv", sep="")
write.csv(res, save, row.names = FALSE)

wtl = c("win", "tie", "loss")
colnames(res) = wtl 

save = paste(FolderResults, "/clp.pdf", sep="")
win.tie.loss.plot(data = res, 
                  names.methods = methods.names, 
                  name.file = save, width = 18, height = 10, 
                  bottom = 2, left = 11, top = 0, right = 1, 
                  size.font = 2.0,wtl = wtl)


}
